# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mora e-commerce admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [orders]
      security: [{ bearerAuth: [] }]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [orders, drafts, abandoned]
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: adminCreateOrder
      tags: [orders]
      security: [{ bearerAuth: [] }]
      summary: Create an order (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [orders]
      security: [{ bearerAuth: [] }]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
    put:
      operationId: adminUpdateOrder
      tags: [orders]
      security: [{ bearerAuth: [] }]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
    delete:
      operationId: adminDeleteOrder
      tags: [orders]
      security: [{ bearerAuth: [] }]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Customers ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DeletedResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            deleted:
              type: boolean
          required: [deleted]
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    ListMeta:
      type: object
      properties:
        total:
          type: integer
      required: [total]

    # ─── Product ──────────────────────────────────────────────────────────────
...
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        customerId: { type: ["string", "null"] }
        email: { type: string }
        status: { type: string }
        financialStatus: { type: string }
        fulfillmentStatus: { type: string }
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        shippingAddress: { type: object }
        lineItems: { type: array, items: { type: object } }
        note: { type: string }
        tags: { type: array, items: { type: string } }
        isDraft: { type: boolean }
        isAbandoned: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, orderNumber, email, status, total]
    OrderInput:
      type: object
      properties:
        email: { type: string }
        shippingAddress: { type: object }
        lineItems: { type: array, items: { type: object } }
        note: { type: string }
      required: [email]
    OrderUpdate:
      type: object
      properties:
        status: { type: string }
        financialStatus: { type: string }
        fulfillmentStatus: { type: string }
        note: { type: string }
    OrderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    OrderResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Order"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Customer ─────────────────────────────────────────────────────────────
...